requirements.txt
pika==1.3.2
redis==5.2.1
psycopg2-binary==2.9.10
flowise==1.0.4
python-dotenv==1.0.0
